requirements.txt
faiss-cpu
fastapi==0.135.2
groq==1.1.2
numpy==2.4.3
pydantic==2.12.5
python-dotenv==1.2.2
Requests==2.33.0
scikit_learn==1.8.0
sentence_transformers==5.3.0
yfinance==1.2.0
uvicorn
